remove PaintNodePrivate.hpp
